feat: rename OceanDonutModel to OceanChartModel
