+++ ocean-components/src/main/java/br/com/useblu/oceands/bindingadapters/BindingAdapterChartCard.kt
@@ -3,12 +3,12 @@ package br.com.useblu.oceands.bindingadapters
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import br.com.useblu.oceands.adapter.OceanChartLegendAdapter
-import br.com.useblu.oceands.model.OceanDonutModel
+import br.com.useblu.oceands.model.chart.OceanChartModel
 
 @BindingAdapter("items")
 fun setChartCardAdapter(
     recyclerView: RecyclerView,
-    model: OceanDonutModel?
+    model: OceanChartModel?
 ) {
     model?.let {
         val oceanShortcutsListAdapter = OceanChartLegendAdapter(it)
