--- ocean-components/src/main/java/br/com/useblu/oceands/model/OceanDonutModel.kt
@@ -1,9 +1,0 @@
-package br.com.useblu.oceands.model
-
-data class OceanDonutModel(
-    val title: String = "",
-    val label: String = "",
-    val items: List<OceanDonutItem> = emptyList(),
-    val onItemSelected: (OceanDonutItem) -> Unit = {},
-    val onNothingSelected: () -> Unit = {}
-)
