+++ ocean-components/src/main/res/layout/ocean_donut.xml
@@ -7,11 +7,11 @@
 
         <import type="java.util.List" />
 
-        <import type="br.com.useblu.oceands.model.OceanDonutModel" />
+        <import type="br.com.useblu.oceands.model.chart.OceanChartModel" />
 
         <variable
             name="model"
-            type="OceanDonutModel" />
+            type="br.com.useblu.oceands.model.chart.OceanChartModel" />
 
     </data>
